requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
yadisk==3.4.0
supabase==1.1.1
pandas==2.1.4
matplotlib==3.8.2
openpyxl==3.1.2
xlsxwriter==3.1.9
requests